Consolidated network-data with observations in web services since they now share an endpoint constant.
